refactor(ironfish): s/useRawTxFixture/buildRawTransaction
