+++ ironfish/src/testUtilities/fixtures.ts
@@ -15,6 +15,7 @@ import { SerializedTransaction, Transaction } from '../primitives/transaction'
 import { IJSON } from '../serde'
 import { Account, AccountValue, Wallet } from '../wallet'
 import { WorkerPool } from '../workerPool/pool'
+import { buildRawTransaction } from './helpers/transaction'
 import { getCurrentTestPath } from './utils'
 
 const FIXTURE_FOLDER = '__fixtures__'
@@ -273,7 +274,7 @@ export async function useBlockWithRawTxFixture(
   sequence: number,
 ): Promise<Block> {
   const generate = async () => {
-    const transaction = await useRawTxFixture(
+    const transaction = await buildRawTransaction(
       chain,
       pool,
       sender,
@@ -289,44 +290,6 @@ export async function useBlockWithRawTxFixture(
   }
 
   return useBlockFixture(chain, generate)
-}
-
-export async function useRawTxFixture(
-  chain: Blockchain,
-  pool: WorkerPool,
-  sender: Account,
-  notesToSpend: NoteEncrypted[],
-  receives: { publicAddress: string; amount: bigint; memo: string; assetIdentifier: Buffer }[],
-  mints: MintDescription[],
-  burns: BurnDescription[],
-): Promise<Transaction> {
-  const spends = await Promise.all(
-    notesToSpend.map(async (n) => {
-      const note = n.decryptNoteForOwner(sender.incomingViewKey)
-      Assert.isNotUndefined(note)
-      const treeIndex = await chain.notes.leavesIndex.get(n.merkleHash())
-      Assert.isNotUndefined(treeIndex)
-      const witness = await chain.notes.witness(treeIndex)
-      Assert.isNotNull(witness)
-
-      return {
-        note,
-        treeSize: witness.treeSize(),
-        authPath: witness.authenticationPath,
-        rootHash: witness.rootHash,
-      }
-    }),
-  )
-
-  return pool.createTransaction(
-    sender.spendingKey,
-    spends,
-    receives,
-    mints,
-    burns,
-    BigInt(0),
-    0,
-  )
 }
 
 export async function useMinersTxFixture(
